Adding PreventManipulationAccessMixin, and changes Views.
PreventManipulationAccessMixin is a new AccessMixin, which
will act as LoginRequiredMixin and restrict access on the
view only for who created the object.

Changes View configuration, code refactoring, enhancing docs
and new mixins to handle the Views, PreventManipulationAccessMixin
will only used with Update and Delete View, all the other Views
have a LoginRequired mixins which prevents Anonymous request to the
website.